third_party/blink/web_tests/external/wpt/css/cssom-view/scrollIntoView-iframes-expected.txt

This is a testharness.js-based test.
[FAIL] scrollIntoView in cross-origin iframe can't scroll in parent window
  assert_equals: scrollTop expected 0 but got 1200
Harness: the test ran to completion.
